requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
Automat==25.4.16
beautifulsoup4==4.14.3
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
constantly==23.10.4
cryptography==46.0.3
cssselect==1.3.0
dataclasses-json==0.6.7
ddgs==9.10.0
debugpy==1.8.19
decorator==5.2.1
distro==1.9.0
et_xmlfile==2.0.0
executing==2.2.1
fake-useragent==2.2.0
filelock==3.20.1
frozenlist==1.8.0
fuzzywuzzy==0.18.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.47.0
google-auth-oauthlib==1.2.4
google_search_results==2.4.2
greenlet==3.3.0
groq==1.0.0
gspread==6.2.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
itemadapter==0.13.0
itemloaders==1.3.2
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
langchain==1.2.4
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-tavily==0.2.16
langchain-text-splitters==1.1.0
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
Levenshtein==0.27.3
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.14.0
nest-asyncio==1.6.0
nltk==3.9.2
numpy==2.4.0
oauthlib==3.3.1
openpyxl==3.1.5
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
parsel==1.10.0
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
primp==0.15.0
prompt_toolkit==3.0.52
propcache==0.4.1
Protego==0.5.0
protobuf==6.33.2
psutil==7.2.1
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
PyDispatcher==2.0.7
Pygments==2.19.2
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-Levenshtein==0.27.3
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
queuelib==1.8.0
RapidFuzz==3.14.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
Scrapy==2.7.1
serpapi==0.1.5
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8.3
SQLAlchemy==2.0.45
stack-data==0.6.3
streamlit==1.52.2
tavily-python==0.7.19
tenacity==9.1.2
tiktoken==0.12.0
tldextract==5.3.1
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
Twisted==25.5.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uuid_utils==0.13.0
w3lib==2.3.1
watchdog==6.0.0
wcwidth==0.2.14
xxhash==3.6.0
yarl==1.22.0
zope.interface==8.1.1
zstandard==0.25.0
duckduckgo_search